Remade input handling by creating input manager and reciever
I made an Input Manager that sends all inputs to every reciever of inputs. Recievers are part of an abstract class called InputReciever, and they then check what the input is called and compares it to their manually made commands. If they have the same name their values are updated and can be used